updates import statements of classes touched by the change: "renames package selection to tokenizers"
