+++ src/main/java/com/vesperin/text/Selection.java
@@ -9,8 +9,8 @@ import com.vesperin.base.locations.Locations;
 import com.vesperin.base.locators.ProgramUnitLocation;
 import com.vesperin.base.locators.UnitLocation;
 import com.vesperin.text.nouns.Noun;
-import com.vesperin.text.selection.ASTNodeWordsTokenizer;
-import com.vesperin.text.selection.WordsTokenizer;
+import com.vesperin.text.tokenizers.ASTNodeWordsTokenizer;
+import com.vesperin.text.tokenizers.WordsTokenizer;
 import com.vesperin.text.spelling.SpellCorrector;
 import com.vesperin.text.spelling.StopWords;
 import com.vesperin.text.utils.Jamas;
